Refactors the ZoneGraph class into it's own module
