+++ tests/test_template.py
@@ -14,7 +14,6 @@ from archetypal.template import (
     YearSchedule,
     ZoneConstructionSet,
     ZoneDefinition,
-    ZoneGraph,
     calc_simple_glazing,
 )
 from archetypal.template.umi_base import UniqueName, load_json_objects
@@ -1537,7 +1536,6 @@ class TestZoneLoad:
         import json
 
         from archetypal.template import ZoneLoad
-        from archetypal.utils import reduce
 
         filename = "tests/input_data/umi_samples/BostonTemplateLibrary_2.json"
         with open(filename, "r") as f:
@@ -2397,121 +2395,6 @@ class TestBuildingTemplate:
             bt:
         """
         assert bt
-
-
-class TestZoneGraph:
-    """Series of tests for the :class:`ZoneGraph` class"""
-
-    def test_traverse_graph(self, small_office):
-        """
-        Args:
-            small_office:
-        """
-
-        idf = small_office
-
-        G = ZoneGraph.from_idf(idf, log_adj_report=False)
-
-        assert G
-
-    @pytest.fixture(scope="module")
-    def G(self, config, small_office):
-        """
-        Args:
-            config:
-            small_office:
-        """
-
-        idf = small_office
-        yield ZoneGraph.from_idf(idf)
-
-    @pytest.mark.parametrize("adj_report", [True, False])
-    def test_graph(self, small_office, adj_report):
-        """Test the creation of a BuildingTemplate zone graph. Parametrize the
-        creation of the adjacency report
-
-        Args:
-            small_office:
-            adj_report:
-        """
-        import networkx as nx
-
-        idf = small_office
-
-        G1 = ZoneGraph.from_idf(idf, log_adj_report=adj_report)
-        assert not nx.is_empty(G1)
-        from eppy.bunch_subclass import EpBunch
-
-        assert isinstance(
-            G1.nodes["Sp-Attic Sys-0 Flr-2 Sch-- undefined - HPlcmt-core ZN"][
-                "epbunch"
-            ],
-            EpBunch,
-        )
-
-    def test_graph_info(self, G):
-        """test the info method on a ZoneGraph
-
-        Args:
-            G:
-        """
-        G.info()
-
-    def test_viewgraph2d(self, G):
-        """test the visualization of the zonegraph in 2d
-
-        Args:
-            G:
-        """
-        import networkx as nx
-
-        G.plot_graph2d(
-            nx.layout.circular_layout,
-            (1),
-            font_color="w",
-            legend=True,
-            font_size=8,
-            color_nodes="core",
-            node_labels_to_integers=True,
-            plt_style="seaborn",
-            save=False,
-            show=False,
-            filename="test",
-        )
-
-    @pytest.mark.parametrize("annotate", [True, "Name", ("core", None)])
-    def test_viewgraph3d(self, G, annotate):
-        """test the visualization of the zonegraph in 3d
-
-        Args:
-            G:
-            annotate:
-        """
-        G.plot_graph3d(
-            annotate=annotate,
-            axis_off=True,
-            save=False,
-            show=False,
-        )
-
-    def test_core_graph(self, G):
-        """
-        Args:
-            G:
-        """
-        H = G.core_graph
-
-        assert len(H) == 1  # assert G has no nodes since Warehouse does not have a
-        # core zone
-
-    def test_perim_graph(self, G):
-        """
-        Args:
-            G:
-        """
-        H = G.perim_graph
-
-        assert len(H) > 0  # assert G has at least one node
 
 
 class TestUniqueName(object):
